+++ Assets/PatchKit Patcher/Scripts/AppData/Remote/WebRequestWraps/UnityHttpClient.cs
@@ -44,11 +44,11 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
 
                 try
                 {
-                    // HACK: because WWW is broken and sometimes just does not return STATUS in responseHeaders we are returning status code 200 (we can assume that status code is not an error since www.error is null or empty).
                     if (!www.responseHeaders.ContainsKey("STATUS"))
                     {
-                        DebugLogger.LogWarning("Response headers doesn't contain status information. Since WWW marks response as one without errors, status code is set to 200 (OK).");
-                        result.StatusCode = 200;
+                        // Based on tests, if response doesn't contain status it has probably timed out.
+                        DebugLogger.Log(string.Format("Response is missing STATUS header. Marking it as timed out."));
+                        result.WasTimeout = true;
                     }
                     else
                     {
@@ -62,10 +62,9 @@ namespace PatchKit.Unity.Patcher.AppData.Remote
                         }
                         else
                         {
-                            // HACK: Again, we can't parse the status code (it might be in some different format) - so we simply set it to 200.
-                            DebugLogger.LogWarning(
-                                "Unable to parse status code. Since WWW marks response as one without errors, status code is set to 200 (OK).");
-                            result.StatusCode = 200;
+                            // Based on tests, if response contains invalid status it has probably timed out.
+                            DebugLogger.Log(string.Format("Response has invalid status - {0}. Marking it as timed out.", status));
+                            result.WasTimeout = true;
                         }
                     }